Decompose MCP tool implementations
