+++ brain/app/mcp/tools/__init__.py
@@ -1,0 +1,1 @@
+"""Focused MCP tool implementations used by the protocol server."""
